Change mqtt light service to work with new interfaces
