+++ model/events/abstract_event.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class Event(ABC):
+    """
+    Abstract class for events.
+    """
